김은비 210721 회원정보CRUD 토큰 수정중

+++ backend/src/main/java/com/ssafy/api/controller/UserController.java
@@ -6,6 +6,8 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -54,7 +56,7 @@ public class UserController {
 			@RequestBody @ApiParam(value="회원가입 정보", required = true) UserRegisterPostReq registerInfo) {
 		
 		
-		System.out.println("들어왔다아아앙");
+		System.out.println("회원가입");
 		//임의로 리턴된 User 인스턴스. 현재 코드는 회원 가입 성공 여부만 판단하기 때문에 굳이 Insert 된 유저 정보를 응답하지 않음.
 		User user = userService.createUser(registerInfo);
 		if(user == null)
@@ -84,12 +86,13 @@ public class UserController {
 		SsafyUserDetails userDetails = (SsafyUserDetails)authentication.getDetails();
 		String userId = userDetails.getUsername();
 		User user = userService.getUserByUserId(userId);
-		System.out.println(user.getPassword());
+		System.out.println(user.toString());
+//		System.out.println(user.getPassword());
 		
 		return ResponseEntity.status(200).body(UserRes.of(user));
 	}
 
-	@DeleteMapping("/delete")
+	@DeleteMapping("/{userId}")
 	@ApiOperation(value = "회원 탈퇴", notes = "로그인한 회원 본인의 정보를 삭제한다.") 
 	@ApiResponses({
 		@ApiResponse(code = 200, message = "성공"),
@@ -97,21 +100,24 @@ public class UserController {
 		@ApiResponse(code = 404, message = "사용자 없음"),
 		@ApiResponse(code = 500, message = "서버 오류")
 	})
-	public ResponseEntity<UserRes> deleteUserInfo(@ApiIgnore Authentication authentication) {
+	public ResponseEntity<? extends BaseResponseBody> deleteUserInfo(@ApiIgnore Authentication authentication, @PathVariable("userId") String userId) {
 		/**
 		 * 요청 헤더 액세스 토큰이 포함된 경우에만 실행되는 인증 처리이후, 리턴되는 인증 정보 객체(authentication) 통해서 요청한 유저 식별.
 		 * 액세스 토큰이 없이 요청하는 경우, 403 에러({"error": "Forbidden", "message": "Access Denied"}) 발생.
 		 */
+		System.out.println("회원 탈퇴");
 		SsafyUserDetails userDetails = (SsafyUserDetails)authentication.getDetails();
-		String userId = userDetails.getUsername();
-		User user = userService.getUserByUserId(userId);
-		Long Id = user.getId();
-		userService.deleteById(Id);
-		
-		return ResponseEntity.status(200).body(UserRes.of(user));
+		String getUserId = userDetails.getUsername();
+		if(getUserId.equals(userId)) {
+			User user = userService.getUserByUserId(userId);
+			Long Id = user.getId();
+			userService.deleteById(Id);
+			return ResponseEntity.status(200).body(BaseResponseBody.of(200, "Success"));
+		}
+		return ResponseEntity.status(500).body(BaseResponseBody.of(500, "fail"));
 	}
 	
-	@PutMapping("/modify")
+	@PatchMapping("/{userId}")
 	@ApiOperation(value = "회원 정보 수정", notes = "로그인한 회원 본인의 정보를 수정한다.") 
 	@ApiResponses({
 		@ApiResponse(code = 200, message = "성공"),
@@ -119,22 +125,31 @@ public class UserController {
 		@ApiResponse(code = 404, message = "사용자 없음"),
 		@ApiResponse(code = 500, message = "서버 오류")
 	})
-	public ResponseEntity<UserRes> modifyUserInfo(@ApiIgnore Authentication authentication,
+	public ResponseEntity<? extends BaseResponseBody> modifyUserInfo(@ApiIgnore Authentication authentication,@PathVariable("userId") String userId,
 			@RequestBody @ApiParam(value="회원 정보 수정", required = true) UserModifyPutReq modifyInfo) {
 		/**
 		 * 요청 헤더 액세스 토큰이 포함된 경우에만 실행되는 인증 처리이후, 리턴되는 인증 정보 객체(authentication) 통해서 요청한 유저 식별.
 		 * 액세스 토큰이 없이 요청하는 경우, 403 에러({"error": "Forbidden", "message": "Access Denied"}) 발생.
 		 */
+		
+		System.out.println("회원정보수정");
+		
 		SsafyUserDetails userDetails = (SsafyUserDetails)authentication.getDetails();
-		String userId = userDetails.getUsername();
-		User user = userService.getUserByUserId(userId);
-		Long Id = user.getId();
+		String getUserId = userDetails.getUsername();
 		
-		User modify_user = userService.modifyUser(modifyInfo, userId, Id);
+		System.out.println(getUserId);
+		System.out.println(userId);
 		
+		if(getUserId.equals(userId)) {
+			User getUser = userService.getUserByUserId(userId);
+			Long id = getUser.getId();
+			User modifyUser = userService.modifyUser(modifyInfo, userId, id);
+			
+			return ResponseEntity.status(200).body(BaseResponseBody.of(200, "Success"));
+		}
 		
+		return ResponseEntity.status(500).body(BaseResponseBody.of(500, "수정 불가"));
 		
-		return ResponseEntity.status(200).body(UserRes.of(modify_user));
 	}
